requirements.txt
streamlit
transformers
PyPDF2
python-docx
torch
